made edits to testing loop in main.cpp to delete random entries with ids from test data. made edits in formatting to linkedlist's printList. made changes in to deleteEntry in hashtable.cpp to only call delete node if id received is valid

+++ hashtable.cpp
@@ -60,11 +60,15 @@ string Hashtable::getData(int id){
 
 bool Hashtable::removeEntry(int id){
     bool removed = false;
-    int position = hash(id);
 
-    if (hashtable[position].deleteNode(id)){
-        removed = true;
-        count--;
+    //if id is not valid, don't even bother calling deleteNode
+    if (id > 0){
+        int position = hash(id);
+
+        if (hashtable[position].deleteNode(id)){
+            removed = true;
+            count--;
+        }
     }
     return removed;
 }
